Revisi K03 & K04

+++ database/migrations/2025_05_13_005111_create_detail_pengawasan_peralatan_table.php
@@ -13,23 +13,25 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('detail_pengawasan_peralatan', function (Blueprint $table) {
-            $table->id();
+        if (!Schema::hasTable('detail_pengawasan_peralatan')){
+            Schema::create('detail_pengawasan_peralatan', function (Blueprint $table) {
+                $table->id();
 
-            $table->unsignedBigInteger('pengawasan_peralatan_id');
-            $table->string('nama_varian_peralatan');
-            $table->string('nama_sub_varian_peralatan');
-            $table->string('merk_peralatan')->nullable();
-            $table->integer('jumlah_unit')->default(0);
-            $table->string('surat_keterangan_k3')->nullable();
-            $table->string('bukti_kepemilikan')->nullable();
-            $table->enum('pencatatan_simpk', ['Sudah', 'Belum'])->default('Belum');
-            $table->string('nomor_registrasi_simpk')->nullable();
+                $table->unsignedBigInteger('pengawasan_peralatan_id');
+                $table->string('nama_varian_peralatan');
+                $table->string('nama_sub_varian_peralatan');
+                $table->string('merk_peralatan')->nullable();
+                $table->integer('jumlah_unit')->default(0);
+                $table->string('surat_keterangan_k3')->nullable();
+                $table->string('bukti_kepemilikan')->nullable();
+                $table->enum('pencatatan_simpk', ['Sudah', 'Belum'])->default('Belum');
+                $table->string('nomor_registrasi_simpk')->nullable();
 
-            $table->timestamps();
+                $table->timestamps();
 
-            $table->foreign('pengawasan_peralatan_id')->references('id')->on('pengawasan_peralatan')->onDelete('cascade');
-        });
+                $table->foreign('pengawasan_peralatan_id')->references('id')->on('pengawasan_peralatan')->onDelete('cascade');
+            });
+         }
     }
 
     /**
